Extract Alb and ApiGW modules from the generated mod
These modules are stable, and we can add improvements
without being tied to the generator.
