+++ aws_lambda_events/Cargo.toml
@@ -71,4 +71,4 @@ ses = []
 sns = []
 sqs = []
 streams = []
-generated = ["activemq", "alb", "apigw", "appsync", "autoscaling", "chime_bot", "clientvpn", "cloudwatch_logs", "code_commit", "codebuild", "codedeploy", "codepipeline", "codepipeline_cloudwatch", "codepipeline_job", "cognito", "config", "connect", "ecr_scan", "firehose", "iot", "iot_1_click", "iot_button", "kafka", "kinesis", "kinesis_analytics", "lex", "rabbitmq", "s3", "s3_batch_job", "ses", "sns", "sqs", "streams"]
+generated = ["activemq", "appsync", "autoscaling", "chime_bot", "clientvpn", "cloudwatch_logs", "code_commit", "codebuild", "codedeploy", "codepipeline", "codepipeline_cloudwatch", "codepipeline_job", "cognito", "config", "connect", "ecr_scan", "firehose", "iot", "iot_1_click", "iot_button", "kafka", "kinesis", "kinesis_analytics", "lex", "rabbitmq", "s3", "s3_batch_job", "ses", "sns", "sqs", "streams"]
