+++ aws_lambda_events/src/alb/mod.rs
@@ -1,4 +1,4 @@
-use super::super::encodings::Body;
+use super::encodings::Body;
 use crate::custom_serde::*;
 use http::{HeaderMap, Method};
 use std::collections::HashMap;
@@ -74,7 +74,9 @@ mod test {
     #[test]
     #[cfg(feature = "alb")]
     fn example_alb_lambda_target_request_headers_only() {
-        let data = include_bytes!("fixtures/example-alb-lambda-target-request-headers-only.json");
+        let data = include_bytes!(
+            "../generated/fixtures/example-alb-lambda-target-request-headers-only.json"
+        );
         let parsed: AlbTargetGroupRequest = serde_json::from_slice(data).unwrap();
         let output: String = serde_json::to_string(&parsed).unwrap();
         let reparsed: AlbTargetGroupRequest = serde_json::from_slice(output.as_bytes()).unwrap();
@@ -84,8 +86,9 @@ mod test {
     #[test]
     #[cfg(feature = "alb")]
     fn example_alb_lambda_target_request_multivalue_headers() {
-        let data =
-            include_bytes!("fixtures/example-alb-lambda-target-request-multivalue-headers.json");
+        let data = include_bytes!(
+            "../generated/fixtures/example-alb-lambda-target-request-multivalue-headers.json"
+        );
         let parsed: AlbTargetGroupRequest = serde_json::from_slice(data).unwrap();
         let output: String = serde_json::to_string(&parsed).unwrap();
         let reparsed: AlbTargetGroupRequest = serde_json::from_slice(output.as_bytes()).unwrap();
@@ -95,7 +98,7 @@ mod test {
     #[test]
     #[cfg(feature = "alb")]
     fn example_alb_lambda_target_response() {
-        let data = include_bytes!("fixtures/example-alb-lambda-target-response.json");
+        let data = include_bytes!("../generated/fixtures/example-alb-lambda-target-response.json");
         let parsed: AlbTargetGroupResponse = serde_json::from_slice(data).unwrap();
         let output: String = serde_json::to_string(&parsed).unwrap();
         let reparsed: AlbTargetGroupResponse = serde_json::from_slice(output.as_bytes()).unwrap();
